Added actual game functionality
Guesses can now be made and clues are given back. The user is given as many guesses as needed until they get the correct word. No time limits, restrictions on guessed words. (except that it must be 5 letters long) Words can be repeated, which will be fixed at some point.